+++ kwcoco/coco_dataset.py
@@ -1170,7 +1170,9 @@ class MixinCocoExtras(object):
             verbose = kwargs.get('verbose', 1)
             res = parse.parse('vidshapes{num_videos:d}', key)
             if res is None:
-                res = parse.parse('vidshapes{num_videos:d}{suffix}', key)
+                res = parse.parse('vidshapes{num_videos:d}-{suffix}', key)
+            if res is None:
+                res = parse.parse('vidshapes-{suffix}', key)
             """
             The rule is that the suffix will be split by the '-' character
             and any registered pattern or alias will impact the kwargs
@@ -1180,7 +1182,8 @@ class MixinCocoExtras(object):
             if verbose > 3:
                 print('res = {!r}'.format(res))
             if res:
-                kwargs['num_videos'] = int(res.named['num_videos'])
+                if 'num_videos' in res.named:
+                    kwargs['num_videos'] = int(res.named['num_videos'])
                 if 'suffix' in res.named:
                     suff_parts = [p for p in res.named['suffix'].split('-') if p]
             if verbose > 3:
@@ -3265,7 +3268,7 @@ class MixinCocoDraw(object):
         canvas = dets.draw_on(canvas)
         return canvas
 
-    def show_image(self, gid=None, aids=None, aid=None, channels=None, **kwargs):
+    def show_image(self, gid=None, aids=None, aid=None, channels=None, setlim=None, **kwargs):
         """
         Use matplotlib to show an image with annotations overlaid
 
@@ -3274,6 +3277,7 @@ class MixinCocoDraw(object):
             aids (list): aids to highlight within the image
             aid (int): a specific aid to focus on. If gid is not give,
                 look up gid based on this aid.
+            setlim (None | str): if 'image' sets the limit to the image extent
             **kwargs:
                 show_annots, show_aid, show_catname, show_kpname,
                 show_segmentation, title, show_gid, show_filename,
@@ -3568,6 +3572,16 @@ class MixinCocoDraw(object):
             if keypoints:
                 xs, ys = np.vstack(keypoints).T
                 ax.plot(xs, ys, 'bo')
+
+        if setlim:
+            if not isinstance(setlim, str):
+                setlim = 'image'
+
+            if setlim == 'image':
+                ax.set_xlim(0, img['width'])
+                ax.set_ylim(img['height'], 0)
+            else:
+                raise NotImplementedError(setlim)
 
         return ax